Fix a semantic issue - KIND OF, not really. See msg.
I accidentally was initializing a shared pointer to point to tvec, when what we wanted was a shared pointer with a new value of the same type as tvec
However, current code is also wrong and will probably segfault when summonForces is called. Because this just makes a shared ptr of the same type, but with a nullptr value.
Anyway, what I basically need to do is "extract" the decltype from the sharedptr, but I'm going to dinner right now so will do this at soonest when I get back.

(I think I can extract the decltype by calling decltype on the dereferenced tvec, then pass that to make_shared)