Some draw routine refactoring, see whole message
-Commented out some of the sound play commands we don't have yet
(and that should be compartmentalized out of draw functions)

-Started refactoring drawMeters. Really drawMeters should be a
game-rules/HUD-level resolution, and should only poll
players/characters for information, not give them draw routines for
meters. My thoughts on how to refactor this is as follows:

Make a small mini-class for counters (e.g. meters) in general. This
probably inherits from vector<int>, but also includes a color, specified
as RGB values, and a target max value (For display purposes, so it can
be shrunk or grown as demanded by the available space on-screen and
scaled to still be accurate in terms of ratios)